Adjust the tests
